Nets with the name of a bus were not handled properly.
More specificaly, when a net had a name such as 'dec_br_taken_e1[0]', the brackets were stripped in a regex and then the net was never found again, since it was not a bus but a regular wire!

Note: the '\' in front of the name means that the string is containing special characters, such as brackets. It acts as a global escaping character for the String in Verilog syntax.